+++ geompp_python/geompp/__init__.py
@@ -31,6 +31,9 @@ Free functions
     average(points)
     signed_area(points[, plane])
     centroid(points[, plane])
+    find_extreme_points(polygon, line)
+    distance_to(polygon, line)
+    tangents_to(polygon, point_or_polygon)
 
 Precision
 ---------
@@ -110,6 +113,9 @@ from ._geompp import (  # noqa: F401
     ExtremePoints3D,
     find_extreme_points,
     distance_to,
+    PolygonTangents2D,
+    PolygonTangents3D,
+    tangents_to,
 )
 
 __version__ = "1.0.0"
@@ -130,5 +136,6 @@ __all__ = [
     "has_intersections", "find_intersections", "convex_hull",
     "CoordinateFrame", "principal_axes", "principal_normal", "principal_direction",
     "ExtremePoints2D", "ExtremePoints3D", "find_extreme_points", "distance_to",
+    "PolygonTangents2D", "PolygonTangents3D", "tangents_to",
     "ProjectionType", "View2D",
 ]